remove last of glide
