+++ app/proguard-rules.pro
@@ -27,14 +27,6 @@
 -dontwarn javax.annotation.**
 -dontwarn retrofit2.Platform$Java8
 
-# Glide specific rules #
-# https://github.com/bumptech/glide
--keep public class * implements com.bumptech.glide.module.GlideModule
--keep public class * extends com.bumptech.glide.AppGlideModule
--keep public enum com.bumptech.glide.load.resource.bitmap.ImageHeaderParser$** {
-    **[] $VALUES;
-    public *;
-}
 
 # RxJava 1.1.0
 -dontwarn sun.misc.**
