+++ app/src/main/java/eu/kanade/tachiyomi/ui/manga/MangaHeaderHolder.kt
@@ -9,6 +9,7 @@ import android.view.ViewGroup
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.core.content.ContextCompat
 import androidx.core.graphics.ColorUtils
+import coil.api.clear
 import coil.api.load
 import coil.api.loadAny
 import coil.request.CachePolicy
@@ -309,16 +310,12 @@ class MangaHeaderHolder(
         }
     }
 
-    fun updateCover(manga: Manga) {
-       // if (!isCached(manga)) return
+    fun updateCover(manga: Manga, forceUpdate:Boolean = false) {
+       if(!isCached(manga) && !forceUpdate) return
+        manga_cover.clear()
+        backdrop.clear()
         manga_cover.loadAny(manga)
         backdrop.loadAny(manga)
-    /*    GlideApp.with(view.context).load(manga).diskCacheStrategy(DiskCacheStrategy.RESOURCE)
-            .signature(ObjectKey(MangaImpl.getLastCoverFetch(manga.id!!).toString()))
-            .into(manga_cover)
-        GlideApp.with(view.context).load(manga).diskCacheStrategy(DiskCacheStrategy.RESOURCE)
-            .signature(ObjectKey(MangaImpl.getLastCoverFetch(manga.id!!).toString())).centerCrop()
-            .transition(DrawableTransitionOptions.withCrossFade()).into(backdrop)*/
     }
 
     private fun isCached(manga: Manga): Boolean {
