+++ app/src/main/java/eu/kanade/tachiyomi/ui/migration/MangaHolder.kt
@@ -3,9 +3,7 @@ package eu.kanade.tachiyomi.ui.migration
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import coil.api.clear
-import coil.api.load
 import coil.api.loadAny
-import com.bumptech.glide.load.engine.DiskCacheStrategy
 import eu.davidea.flexibleadapter.FlexibleAdapter
 import eu.davidea.flexibleadapter.items.IFlexible
 import eu.kanade.tachiyomi.ui.base.holder.BaseFlexibleViewHolder
