+++ app/src/main/java/eu/kanade/tachiyomi/ui/migration/manga/process/MigrationProcessHolder.kt
@@ -4,10 +4,7 @@ import android.view.View
 import androidx.appcompat.widget.PopupMenu
 import androidx.constraintlayout.widget.ConstraintLayout
 import coil.Coil
-import coil.api.clear
 import coil.request.LoadRequest
-import com.bumptech.glide.load.engine.DiskCacheStrategy
-import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
 import eu.kanade.tachiyomi.R
 import eu.kanade.tachiyomi.data.database.DatabaseHelper
 import eu.kanade.tachiyomi.data.database.models.Manga
@@ -23,8 +20,6 @@ import eu.kanade.tachiyomi.util.view.setVectorCompat
 import eu.kanade.tachiyomi.util.view.visible
 import eu.kanade.tachiyomi.util.view.withFadeTransaction
 import eu.kanade.tachiyomi.widget.CoverViewTarget
-import eu.kanade.tachiyomi.widget.StateImageViewTarget
-import kotlinx.android.synthetic.main.manga_grid_item.*
 import kotlinx.android.synthetic.main.manga_grid_item.view.*
 import kotlinx.android.synthetic.main.migration_process_item.*
 import kotlinx.android.synthetic.main.unread_download_badge.view.*
@@ -138,18 +133,9 @@ class MigrationProcessHolder(
         (layoutParams as ConstraintLayout.LayoutParams).verticalBias = 1f
         progress.gone()
 
-            val request = LoadRequest.Builder(view.context).data(manga)
-                .target(CoverViewTarget(cover_thumbnail, progress)).build()
-            Coil.imageLoader(view.context).execute(request)
-
-       /* GlideApp.with(view.context.applicationContext).load(manga).apply {
-            diskCacheStrategy(DiskCacheStrategy.RESOURCE)
-            if (isTo) {
-                transition(DrawableTransitionOptions.withCrossFade())
-                    .into(StateImageViewTarget(cover_thumbnail, progress))
-            } else
-                into(cover_thumbnail)
-        }*/
+        val request = LoadRequest.Builder(view.context).data(manga)
+            .target(CoverViewTarget(cover_thumbnail, progress)).build()
+        Coil.imageLoader(view.context).execute(request)
 
         compact_title.visible()
         gradient.visible()
@@ -173,11 +159,15 @@ class MigrationProcessHolder(
         val latestChapter = mangaChapters.maxBy { it.chapter_number }?.chapter_number ?: -1f
 
         if (latestChapter > 0f) {
-            subtitle.text = context.getString(R.string.latest_,
-                DecimalFormat("#.#").format(latestChapter))
+            subtitle.text = context.getString(
+                R.string.latest_,
+                DecimalFormat("#.#").format(latestChapter)
+            )
         } else {
-            subtitle.text = context.getString(R.string.latest_,
-                context.getString(R.string.unknown))
+            subtitle.text = context.getString(
+                R.string.latest_,
+                context.getString(R.string.unknown)
+            )
         }
     }
